+++ tools/xdhq/NJS/xdhp.h
@@ -26,7 +26,7 @@ namespace xdhp {
 	SCLNJS_F( Register );
 	SCLNJS_F( Listen );
 	SCLNJS_F( SetLayout );
-	SCLNJS_F( GetContent );
+	SCLNJS_F( GetContents );
 	SCLNJS_F( SetContents );
 	SCLNJS_F( SetCasts );
 }